[Fix] Fix types of format string arguments to match format specifier on non-windows 64bit systems.
git-svn-id: https://source.openmpt.org/svn/openmpt/trunk/OpenMPT@2185 56274372-70c3-4bfc-bfc3-4c3a0b034d27

+++ soundlib/MIDIMacros.cpp
@@ -141,7 +141,7 @@ void MIDIMacroConfig::CreateParameteredMacro(char (&parameteredMacro)[MACRO_LENG
 void MIDIMacroConfig::CreateFixedMacro(char (&fixedMacros)[128][MACRO_LENGTH], fixedMacroType macroType) const
 //------------------------------------------------------------------------------------------------------------
 {
-	for(size_t i = 0; i < 128; i++)
+	for(unsigned int i = 0; i < 128; i++)
 	{
 		switch(macroType)
 		{
